requirements.txt
# File names Start capital and camel case
# function and variable naming style - undercase 
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
category_encoders==2.7.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dacite==1.9.2
dash==3.0.2
debugpy==1.8.7
decorator==5.1.1
defog==0.65.8
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
executing==2.1.0
fastapi==0.112.1
fastjsonschema==2.20.0
Flask==2.2.5
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
fonttools==4.54.1
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
lightgbm==4.6.0
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.0.2
mlxtend==0.23.2
multimethod==1.12
mysql-connector-python==9.0.0
narwhals==1.33.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openai==1.41.0
orjson==3.10.16
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prometheus_client==0.21.0
prompt_toolkit==3.0.47
protobuf==5.29.4
psutil==6.1.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
puremagic==1.28
pwinput==1.0.3
pyarrow==19.0.1
pycaret==3.3.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.8.0
pyod==2.0.4
pyodbc==5.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.2.0
simplejson==3.19.3
six==1.16.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.2
statsmodels==0.14.4
streamlit==1.44.1
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
tsdownsample==0.1.4.1
typeguard==4.4.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
visions==0.7.4
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
xxhash==3.5.0
ydata-profiling==4.16.1
yellowbrick==1.5
zipp==3.21.0


streamlit==1.44.1
pandas==2.1.4